test: move count distinct spill test into memory_limit tests
